add fade transition and automatic resizing
slides can now fade over the previous slide (fades are on by default).
each slide can have its own transition duration. the final transition
(to the clear colour) is the same duration as the first transition (from
the clear colour). images with transparent sections should mostly be
avoided when using fades; fading may appear not as expected.
slides are now scaled to always fill the entire window. the method of
doing so can be chosen per tile. the slide can be stretched to fit the
window (may alter the aspect ratio) or zoomed until filling the window
(retains aspect ratio).